08_deploy_prediction_prophet_test/requirements.txt
flask
flask_cors
quandl
matplotlib
numpy
prophet
pandas
pytrends
google-cloud-firestore
google-cloud-storage
google-cloud-logging
gunicorn